Split out abstract draw-lots-to-union for different selection domains
