+++ app/templates/components/project-geometry-edit/development-site.hbs
@@ -1,24 +1,19 @@
-{{!-- add the development-specific tax-lots layergroup --}}
-{{#map.labs-layers
-  layerGroups=(array taxLots)
-  onLayerClick=(if (eq mode 'lots') (action 'handleLayerClick') (action 'noop'))
-  as |layers|}}
-  {{#layers.tooltip as |tooltip|}}
-    {{tooltip-renderer
-      feature=tooltip.feature
-      template=tooltip.layer.tooltipTemplate}}
-  {{/layers.tooltip}}
-{{/map.labs-layers}}
-
-{{#if (eq mode 'lots')}}
-  {{!-- Ad-hoc source and layer for showing selected lots in lot selection mode --}}
-  {{#map.source options=selectedLotsSource as |source|}}
-    {{source.layer
-      layer=selectedLotsLayer
-      before='boundary_country'
-    }}
-  {{/map.source}}
-{{/if}}
+{{#project-geometry-edit/draw-lots-to-union
+  mode=mode
+  model=model
+  map=map
+  as |selectedLots finalGeometry|}}
+  {{#ember-wormhole to='geometry-type-save-box'}}
+    <button
+      type="button"
+      class="button large green expanded project-save-button"
+      disabled={{if (eq mode 'draw') false (or (not selectedLots.features.length))}}
+      onclick={{action 'save' model finalGeometry}}
+    >
+      Save Development Site
+    </button>
+  {{/ember-wormhole}}
+{{/project-geometry-edit/draw-lots-to-union}}
 
 {{#map.source options=(mapbox-geojson-source model.developmentSite) as |source|}}
   {{source.layer
@@ -26,16 +21,5 @@
     before='boundary_country'
   }}
 {{/map.source}}
-
-{{#ember-wormhole to='geometry-type-save-box'}}
-  <button
-    type="button"
-    class="button large green expanded project-save-button"
-    disabled={{if (eq mode 'draw') false (or (not selectedLots.features.length))}}
-    onclick={{action 'save' model}}
-  >
-    Save Project
-  </button>
-{{/ember-wormhole}}
 
 {{yield}}
