+++ app/templates/components/project-geometry-edit/draw-lots-to-union.hbs
@@ -1,0 +1,23 @@
+{{!-- add the development-specific tax-lots layergroup --}}
+{{#map.labs-layers
+  layerGroups=(array taxLots)
+  onLayerClick=(if (eq mode 'lots') (action 'handleLayerClick') (action 'noop'))
+  as |layers|}}
+  {{#layers.tooltip as |tooltip|}}
+    {{tooltip-renderer
+      feature=tooltip.feature
+      template=tooltip.layer.tooltipTemplate}}
+  {{/layers.tooltip}}
+{{/map.labs-layers}}
+
+{{#if (eq mode 'lots')}}
+  {{!-- Ad-hoc source and layer for showing selected lots in lot selection mode --}}
+  {{#map.source options=selectedLotsSource as |source|}}
+    {{source.layer
+      layer=selectedLotsLayer
+      before='boundary_country'
+    }}
+  {{/map.source}}
+{{/if}}
+
+{{yield selectedLots finalGeometry}}
